Approved by PS Jenkins bot, Mathieu Trudel-Lapierre.

Original authors:
  - Manuel de la Pena <manuel@canonical.com>
  - Nick Dedekind <nicholas.dedekind@gmail.com>
  - Pawel Stolowski <pawel.stolowski@canonical.com>
  - Michal Hruby <michal.mhr@gmail.com>
  - Łukasz 'sil2100' Zemczak <lukasz.zemczak@canonical.com>
  - Marco Trevisan (Treviño) <mail@3v1n0.net>
  - Automatic PS uploader <ps-jenkins@lists.canonical.com>
  - Chris Townsend <christopher.townsend@canonical.com>
  - Andrea Azzarone <azzaronea@gmail.com>
  - Matthieu James <matthieu.james@canonical.com>
  - Brandon Schaefer <brandon.schaefer@canonical.com>
  - Michi Henning <michi.henning@canonical.com>
  - Tigran Gabrielyan <tigrangab@gmail.com>
  - Stephen M. Webb <stephen.webb@canonical.com>
  - Sebastien Bacher <seb128@ubuntu.com>
  - MC Return <mc.return@gmx.net>
  - Didier Roche <didier.roche@canonical.com>
  - Diego Sarmentero <diego.sarmentero@canonical.com>
  - Diego Sarmentero <diego.sarmentero@gmail.com>
  - James Henstridge <james@jamesh.id.au>
  - Christopher Lee <chris.lee@canonical.com>
  - Nick Dedekind <nick.dedekind@canonical.com>
  - Mathieu Trudel-Lapierre <mathieu-tl@ubuntu.com>
  - Martin Pitt <martin.pitt@canonical.com>
  - Chris Townsend <Christopher.Townsend@canonical.com>
  - Paolo Rotolo <paolorotolo@ubuntu.com>

(bzr r3325)

+++ unity-shared/StandaloneWindowManager.cpp
@@ -600,7 +600,7 @@ void StandaloneWindowManager::AddProperties(GVariantBuilder* builder)
   unity::variant::BuilderWrapper wrapper(builder);
   wrapper.add(GetScreenGeometry())
          .add("workspace_count", WorkspaceCount())
-         .add("active_window", GetActiveWindow())
+         .add("active_window", (uint64_t)GetActiveWindow())
          .add("screen_grabbed", IsScreenGrabbed())
          .add("scale_active", IsScaleActive())
          .add("scale_active_for_group", IsScaleActiveForGroup())
